Add NagiosProbes to pilotCommands.py
